backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pymongo==4.6.1
pytest==8.0.0
httpx==0.26.0
python-multipart==0.0.9
tenacity==8.2.3
